requirements.txt
Flask
Flask-Cors
PyMySQL
python-dotenv
gunicorn
openai
